+++ backend/tsconfig.paths.json
@@ -4,7 +4,7 @@
     "paths": {
       "@/*": ["src/*"],
       "@routes/*": ["src/routes/*"],
-      "@middlewares/*": ["src/middlewares/*"],
+      "@aops/*": ["src/aops/*"],
       "@controllers/*": ["src/controllers/*"],
       "@services/*": ["src/services/*"],
       "@models/*": ["src/models/*"],
